fix: `split_nodes_image` on text with exclamation mark
This commit will fix the `split_nodes_image` function such that it
doesn't error out on receiving the text input with exclamation mark.

Assumption before: Check if input node is of type text and has an
exclamation mark, it would have an image, and if not, it's invalid
markdown.

As one can infer from above assumption that it was clearly wrong and
text with exclamation mark not intending to have an image is indeed
valid and should be parsed accordingly.

Assumption after: check if input node is of type text and has an
exclamation mark, search for image. If found, split it accordingly else
return the text as TEXT_TYPE_TEXT